+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [1.0.1] - 2022-07-18
 ### Added
 - Support for compiling shaders in unsaved text editor
 - Added animation on shader compilation
@@ -38,7 +40,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Initial release
 
-[Unreleased]: https://github.com/Devaniti/ShaderInspector/compare/v1.0.0...HEAD
+[Unreleased]: https://github.com/Devaniti/ShaderInspector/compare/v1.0.1...HEAD
+[1.0.1]: https://github.com/Devaniti/ShaderInspector/compare/v1.0.0...v1.0.1
 [1.0.0]: https://github.com/Devaniti/ShaderInspector/compare/v0.0.2...v1.0.0
 [0.0.2]: https://github.com/Devaniti/ShaderInspector/compare/v0.0.1...v0.0.2
 [0.0.1]: https://github.com/Devaniti/ShaderInspector/releases/tag/v0.0.1